# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: DEVIL API
  version: 1.0.0
  description: |
    DEVIL AI Agent Platform API
    
    Phase 1: Control Plane - Mission State Machine, Event Log, Approval Engine, Validation Layer, Tool Registry
  contact:
    name: DEVIL Team
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /validation/pre-execution:
    post:
      operationId: preExecutionValidation
      tags: [validation]
      summary: Pre-execution validation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidationRequest"
      responses:
        "200":
          description: Validation report
  /validation/resume:
    post:
      operationId: resumeValidation
      tags: [validation]
      summary: Resume validation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                missionId:
                  type: string
      responses:
        "200":
          description: Validation report
  /tools:
    get:
      operationId: listTools
      tags: [tools]
      summary: List all tools
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: capability
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Tool list
    post:
      operationId: initializeTools
      tags: [tools]
      summary: Initialize tool registry
      responses:
        "200":
          description: Tools initialized
...
  /tools/check:
    post:
      operationId: checkToolExecution
      tags: [tools]
      summary: Check if tool can be executed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                toolName:
                  type: string
                missionId:
                  type: string
      responses:
        "200":
          description: Check result
...
components:
  schemas:
...
    ValidationRequest:
      type: object
      properties:
        target:
          type: string
          enum: [mission, phase, task]
        targetId:
          type: string
        requestedAction:
          type: string
      required: [target, targetId]
...